Fixed two typos in Odin overview docs.
I found two small typos in the description of the shift operator in Odin's overview doc page.

There was a missing "is" for the signed integer shift part and the unsigned integer shift part said "the it" when "it" was likely intended. 

However, "it" is grammatically ambiguous here so I further changed "it" to be "the left operand" also instead, so that the sentence is no longer ambiguous.