Clarified the motivation of explicit parapoly, etc
I noticed that the `#### Procedures` headers for the explicit and implicit parametric polymorphism sections looked very confusingly identical because the `###` and `####` headers in the Markdown are rendering very similarly and extremely hard to distinguish. They don't even look like subsections of the intended parent sections at all. This caused me to become confused since it seemed like I'd already just read multiple headers that were titled as just `Procedures`.

In the process of cleaning that up by making those headers much easier to distinguish from each other, I also realized that the fact that those procedure sections were even related to the mention of explicit parametric polymorphism (parapoly) wasn't clear at all and didn't flow naturally. I didn't even connect the two in my head until more experimentation with some test code, at which point I realized that the "Procedures" sections were actually probably supposed to be subsections of the larger parapoly explanation.

So, I went through the text and made an effort to make the connections between these sections and the motivations and value of the related features much clearer to the readers. The original text didn't give a strong sense of "why".

Less experienced programmers are very likely to be confused by the original text and most of them would not likely make the connection in their heads unless it is explained more clearly, and so that motivated my suggested clarifications and changes for this part of the text.