Updated copyright year range for the Odin site.
When looking at the website, I randomly noticed that the copyright year number for Odin for Ginger Bill was still "2016-2022".

Thus, I found this footer HTML file and edited it here.

Is this the correct way to fix that for the site? Should I be fixing it?

In any case, I thought you should know, and so I'm submitting this proposed change.