Made new section to help b/c Odin's sparse docs.
Given that Odin currently doesn't have much documentation, I have created an new section at the very end of the Overview page whose purpose is to cover a variety of the most likely ways that the user can find information about how to use Odin above and beyond what actually exists in the docs currently.

Users may not be aware that these useful subdirectories of Odin's source folder exist, especially if they are relatively new to programming. They also may not actually think to use them to try to solve their problems. By including a new section to account for this, Odin can increase the chances of people being able to successfully use the language and decrease the chances of them leaving. It will also make the relative lack of documentation in Odin less intimidating and inspire people to proactively search the source code for what they can use.

This seems like an important section to add. There needs to be something that compensates for Odin's low quantity of documentation and this is my attempt at at least doing something to address that. The advice is based on my general programming knowledge (I am still an Odin newbie in contrast), but it seems like valid advice nonetheless and worth saying.

What do you guys think? I hope you'll consider adding it in.

Anyway, as always, feel free to freely edit what I submit.