Added link to explain language servers, etc.
As an additional amendment to my previous commit, I've also now added a Wikipedia link to help users who are unfamiliar with what a "language server" is understand it.

After that, I decided I might as well add a link to the syntax highlighter page too, although the function of syntax highlighting is far more obvious of course.

Some very new programmers may still randomly pickup Odin, so I figure it can't hurt to make a bit of effort to accommodate them too. After all, Odin has great potential as a cleaner way to learn systems programming than C and without the complexity of Rust, etc.

Oh, and I also just now added a link to the list of editors with Odin code support on the community library page too. That way it will be easier for people to find it when they are trying to get started.