Clarify packages in overview
A repeated occurrence in the #beginner channel of the Odin discord chat is new users being confused about how packages work. This is a clarification intended to answer those questions at the outset. 

Language is intended to stay neutral and simply present how things work, in line with the rest of the overview. Please review and consider merging.