Update faq.md - "is outweighed by" to "outweighs"
I'm not sure if this change is actually correct, because I'm not sure about the actual intent behind the sentence. The next sentence seems to imply the paragraph is critiquing a decision of C as being the wrong decision. But maybe this sentence means to say that despite those flaws this decision was still correct in C; in which case, the original phrasing is correct.